refactor(skills): colocate marketplace installation logic
